test: move nested preload tests into preload.rs
Consolidate all nested preload permutation tests into the existing
preload.rs file instead of a separate nested_preload.rs file.

16 new test functions covering all relation type permutations for
nested (two-level) association preloading:

Passing (7 tests):
- HasMany -> BelongsTo<T>, HasMany -> HasMany (empty leaves)
- HasOne<Option<T>> -> BelongsTo<T>
- BelongsTo<T> -> HasMany, BelongsTo<T> -> BelongsTo<T>,
  BelongsTo<T> -> HasOne<T>, BelongsTo<Option<T>> -> HasMany

Ignored / revealing engine bugs (9 tests):
- HasMany -> HasOne<Option<T>>, HasMany -> HasOne<T>,
  HasMany -> BelongsTo<Option<T>>
- HasOne<Option<T>> -> HasMany, HasOne<T> -> HasMany,
  HasOne<Option<T>> -> HasOne<Option<T>>, HasOne<T> -> HasOne<T>
- BelongsTo<T> -> HasOne<Option<T>>,
  BelongsTo<Option<T>> -> BelongsTo<Option<T>>

The ignored tests panic with type mismatches in the query engine's
input resolution, exposing gaps in the nested preload implementation.
