+++ crates/toasty-driver-integration-suite/src/tests.rs
@@ -24,7 +24,6 @@ pub mod has_many_n_1;
 pub mod has_many_scoped_query;
 pub mod has_one_crud_basic;
 pub mod jiff;
-pub mod nested_preload;
 pub mod one_model_batch_create;
 pub mod one_model_composite_key;
 pub mod one_model_crud;
